project docs.git for 5.0.0 manifests should use tag 5.0.0-docs
The docs for 5.0.0 were incomplete when the initial 5.0.0
tags was created.  Now that the missing content has been delivered
to the r/stx.5.0 branch of docs.git.  A 5.0.0-docs tag has
been created to mark the completed 5.0.0 docs, but in the
docs.git project only.

This update will revise the 5.0.0 manifests to use the
5.0.0-docs tag for docs.git.

Story: 2008927
Task: 42527
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I55691792a54354a434affdd479210499ac3a9ab2