Create stx-openstack manifest
As discussed in [1], as part of the STX-Openstack upversion to Antelope
, we have to decouple the platform and the application builds in
separate manifests, in order to not clash the upversioned packages on
the application side, with the ones on the platform.

The purpose of this change is to introduce this new manifest, containing
the required repositories for a successfull build of the STX-Openstack
application.

For the test plan, despite this change being on the
`starlingx/manifest` repo, I wanted to make sure that all required
repositories for a successfull STX-Openstack build were present,
hence why the test plan is not the usual for this repository.

[1] https://lists.starlingx.io/pipermail/starlingx-discuss/2023-June/014214.html

Test Plan:
PASS: Initialize repo with custom manifest
PASS: Compare STX-Openstack tarball generated by this manifest against
      an official one and make sure no charts or FluxCD manifests are
      missing

On build environment:
  PASS: Download all dependencies (downloader -s -b -B std,rt)
  PASS: Build all packages (build-pkgs -c -a)
  PASS: Build wheels
  PASS: Build all STX-Openstack Docker images
  PASS: Build STX-Openstack tarball
  PASS: Upload / apply STX-Openstack

Story: 2010797
Task: 48249

Change-Id: I58510c120c35c0e894aeeef100df37c8c1c749c9
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>