Remove openstack-armada-app from default manifest
As discussed in [1], as part of the STX-Openstack upversion to Antelope
, we have to decouple the platform and the application builds in
separate manifests, in order to not clash the upversioned packages on
the application side, with the ones on the platform.

After the creation of the STX-Openstack manifest, the
`starlingx/openstack-armada-app` repository can be removed from the
default manifest without affecting the build process. Since the only
package this repo delivers to the ISO is the
`stx-openstack-helm-fluxcd`, and it's not actually required to be
delivered on the platform [2], there's no impact on StarlingX

[1] https://lists.starlingx.io/pipermail/starlingx-discuss/2023-June/014214.html
[2] https://review.opendev.org/c/starlingx/openstack-armada-app/+/889991

Test Plan:
PASS: Run a `repo init` with this change
PASS: Perform a `repo sync` and make sure the openstack-armada-app
      repository does not exist anymore

Story: 2010797
Task: 48695

Change-Id: I7b4ed700d57722a11e39766766ae7849fb22e456
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>