Added apt-ostree to WRCP manifest
Right now, apt-ostree is included by pulling a tarball from opendev.
Opendev maintainers have stated this feature cannot be relied on.
And this method also requires an extra commit when updating
apt-ostree to point to the new tarball.

With this change, the repo will be cloned along with the others.
Another commit will be made on the update repo to point to the
cloned repo for building the pkg.

Test Plan:
pass: repo sync and confirm apt-ostree
      at cgcs-root/stx/git/

Story: 2010867
Task: 52332

Change-Id: Ie143e2b7ec9940fa63a5c1bb6c733ebe0e01f630
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>