Added apt-ostree to stx-openstack manifest
Right now, apt-ostree is included by pulling a tarball from opendev.
Opendev maintainers have stated this feature cannot be relied on.
And this method also requires an extra commit when updating
apt-ostree to point to the new tarball.

With this change, the repo will be cloned along with the others.
Another commit will be made on the update repo to point to the
cloned repo for building the pkg.

Test Plan:
pass: repo sync and confirm apt-ostree
          at cgcs-root/stx/git/

Story: 2010867
Task: 52332

Change-Id: I8c3899470ff9b7f5b0918b78e762720b56cb0b71
Signed-off-by: Scott Little <scott.little@windriver.com>